fix: use constants for analytics sources
