+++ assets/scripts/app/constants.js
@@ -64,3 +64,70 @@ export const RESERVED_URLS = [
 ]
 
 export const URL_RESERVED_PREFIX = '~'
+
+export const DEFAULT_ANALYTICS_SOURCE = 'giz'
+
+const CAPACITIES = {
+  sidewalk: { average: 19000, potential: 19000 },
+  'drive-lane': { average: 1500, potential: 2000 },
+  'bike-lane': { average: 14000, potential: 14000 },
+  scooter: { average: 14000, potential: 14000 },
+  'light-rail': { average: 18000, potential: 20000 },
+  streetcar: { average: 18000, potential: 20000 },
+  'bus-lane': { average: 5000, potential: 8000 },
+  'magic-carpet': { average: 2, potential: 3 }
+}
+
+const CAPACITIES_FOO = {
+  sidewalk: { average: 1500, potential: 1500 },
+  'drive-lane': { average: 1500, potential: 1500 },
+  'bike-lane': { average: 1500, potential: 1500 },
+  scooter: { average: 1500, potential: 1500 },
+  'light-rail': { average: 1500, potential: 1500 },
+  streetcar: { average: 18000, potential: 1500 },
+  'bus-lane': { average: 5000, potential: 8000 },
+  'magic-carpet': { average: 1500, potential: 1500 }
+}
+
+const CAPACITIES_BAR = {
+  sidewalk: { average: 5000, potential: 5000 },
+  'drive-lane': { average: 5000, potential: 5000 },
+  'bike-lane': { average: 5000, potential: 5000 },
+  scooter: { average: 5000, potential: 5000 },
+  'light-rail': { average: 18000, potential: 18000 },
+  streetcar: { average: 18000, potential: 20000 },
+  'bus-lane': { average: 5000, potential: 8000 },
+  'magic-carpet': { average: 2, potential: 3 }
+}
+
+export const SOURCE_LIST = [
+  {
+    label: 'Gesellschaft für Internationale Zusammenarbeit (GIZ)',
+    value: 'giz',
+    capacities: CAPACITIES,
+    citationUrl:
+      'http://www.uncrd.or.jp/content/documents/5594Presentation%203%20-%20Module%201%20-%20Mr.%20Breithaupt.pdf',
+    citationLabel:
+      'Environmentally Sustainable Transport - Main Principles and Impacts',
+    citationSub:
+      ', Manfred Breithaupt, Deutsche Gesellschaft für Internationale Zusammenarbeit (GIZ)'
+  },
+  {
+    label: 'Foo Source',
+    value: 'foo',
+    capacities: CAPACITIES_FOO,
+    citationUrl: 'http://google.com',
+    citationLabel: 'Study of Foo in Public Spaces',
+    citationSub: ', Some more text for Foo goes here'
+  },
+  {
+    label: 'Bar Source',
+    value: 'bar',
+    capacities: CAPACITIES_BAR,
+    citationUrl: 'http://streetmix.net',
+    citationLabel: 'Study of Bar in Public Spaces',
+    citationSub: ', Some more text for Bar goes here'
+  }
+]
+
+export const SOURCE_VALUES = SOURCE_LIST.map((item) => item.value)
